+++ TP4/Worker.java
@@ -7,9 +7,9 @@ import java.util.concurrent.Callable;
  * Task for running the Monte Carlo simulation.
  */
 class Worker implements Callable<Long> {
-    private int numIterations;
+    private long numIterations;
 
-    public Worker(int num) {
+    public Worker(long num) {
         this.numIterations = num;
     }
 
@@ -17,7 +17,7 @@ class Worker implements Callable<Long> {
     public Long call() {
         long circleCount = 0;
         Random prng = new Random();
-        for (int j = 0; j < numIterations; j++) {
+        for (long j = 0; j < numIterations; j++) {
             double x = prng.nextDouble();
             double y = prng.nextDouble();
             if ((x * x + y * y) < 1)
